## 2.0-ALPHA7
### Добавлено
* Сущности:
  * `ProjectEntity`
  * `ExpenseItemEntity`
  * `CashOutDocumentEntity`
  * `InventoryDocumentEntity`
  * `InvoiceInDocumentEntity`
  * `PaymentInDocumentEntity`
  * `InvoiceOutDocumentEntity`
  * `PaymentOutDocumentEntity`
  * `ProcessingDocumentEntity`
  * `InternalOrderDocumentEntity`
  * `CommissionReportInDocumentEntity`
  * `CommissionReportOutDocumentEntity`
* Методы API:
  * `GET /entity/group/`
  * `GET /entity/cashout/`
  * `GET /entity/project/`
  * `GET /entity/inventory/`
  * `GET /entity/invoicein/`
  * `GET /entity/paymentin/`
  * `GET /entity/invoiceout/`
  * `GET /entity/paymentout/`
  * `GET /entity/processing/`
  * `GET /entity/expenseitem/`
  * `GET /entity/internalorder/`
  * `GET /entity/commissionreportin/`
  * `GET /entity/commissionreportout/`
  * `GET /entity/cashout/metadata`
  * `GET /entity/project/metadata`
  * `GET /entity/paymentin/metadata`
  * `GET /entity/paymentout/metadata`
  * `POST /entity/group/`
  * `POST /entity/cashout/`
  * `POST /entity/invoicein/`
  * `POST /entity/inventory/`
  * `POST /entity/paymentin/`
  * `POST /entity/invoiceout/`
  * `POST /entity/paymentout/`
  * `POST /entity/processing/`
  * `POST /entity/expenseitem/`
  * `POST /entity/internalorder/`
  * `POST /entity/commissionreportin/`
  * `POST /entity/commissionreportout/`

+++ src/main/java/com/lognex/api/entities/products/AbstractProductEntity.java
@@ -53,7 +53,7 @@ public abstract class AbstractProductEntity extends MetaEntity implements Produc
     /**
      * Минимальная цена
      */
-    private Integer minPrice;
+    private Long minPrice;
 
     /**
      * Закупочная цена
